etl/requirements.txt
pulsar-client==3.0.0